WPF V0.2.1 (Foundation V.4.1.5)
WPF v0.2.1
-Zurb Foundation update to v4.1.5
-Bugfix: $home_url had no default: home_url('/')
-Bugfix: wpf_breadcrumb() gave a notice error when it processed a page.
-Added a fallback function for wp_nav_menu(). It will only be shown if
no menu is selected and the user has the ability to change menus.
-Changed the name for the modernizr to "custom.modernizr-2.6.2.min.js"
so you can see the modernizr version and minification at a glance.
-Added robot.txt/human.txt
-Implemented as much as possible from html5boilerplate
-First steps of cleaning up functions.php and adding phpdoc comments.
-First steps of using wordpress coding standards (see
http://make.wordpress.org/core/handbook/coding-standards/).
-First steps of making theme more semantic (header.php, footer.php,
searchform.php).
-Added a new function wpf_print_footer_sidebar() for printing widget
(removed logic from the footer template part).
-Some more buxfixes/cleaning that I might have forgotten to mention.

I am in a sort of impass. Not really sure if semantic > non-semantic.
Seeing at how hard it is to manage the css with @media.